System I Parse and Extractor v2/requirements.txt
# Core PDF processing
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.0.1

# AI/ML components
torch>=2.2.0
transformers>=4.35.0
nougat-ocr>=0.1.7
layoutparser[ocr]>=0.3.4
opencv-python>=4.8.0

# Text processing and similarity
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1
scikit-learn>=1.3.0
numpy>=1.24.0

# Table extraction
camelot-py[cv]>=0.11.0
tabula-py>=2.8.0

# Reference parsing
pdfx>=1.3.0
sentence-transformers>=2.2.0

# Utilities
pyyaml>=6.0.0
click>=8.1.0
tqdm>=4.66.0
matplotlib>=3.7.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0 